organize curriculum sync code into submodules
